Removed __call() from generted ActiveRecord classes.
The method gets generated only if a behavior implements the "objectCall" hook.
Otherwise, the logic that was previously in the generated _call() method,
used to deal with virtual column getters, is handled by BaseObject.

This commit should unclutter generated AR classs a bit, and use inheritance
when it makes sense (i.e. no variations in code depending on the model).