Fix archivable behavior when used with unique index.
If a table has unique columns, the resulting archive table has set
UNIQUE indices on these columns.

While this seems like a good idea at the first glance, it's a bug
because you can't delete two records which share the same value in this
colum.

This patch fixes the issue by copying unique indices to indices on the
archive table.

Closes #175.