Fix: versioning tables with inheritance
When wersioning a table with inheritance, the `getOneVersion` method
did not return a "Version" object.

Calling `getOneVersion` on a versioned Novel (instance of Book) would
have returned a Book object, instead of a BookVersion object.

Furthermore, the QuickBuilder was broken for this kind of schema and
needed to be fixed.