When a foreign key is part of the primary key, setting it to null
will break the deletion code in the doSave() function on the base OM.

In doSave(), the ::create()->filterByPrimaryKeys($obj->getPrimaryKeys(false))->delete($con)

will not work, because one of the primary key field will be null.

I do not see any harm in removing this code. Tried it and did not see any problems.

Obviously, I have a composite primary key and part of this primary key is a foreign key from another table.