add 2 test cases to show the problem.
testRemove_CompositePK is what symfony form bind() method does when a relation is removed from the main object.

testMove_CompositePK is another example when we try to override the relation, but I cannot make my code works for it.

Also, I changed the "clone" for an unserialize(serialize()) in the set() method because it was causing a problem, but using the same trick in the remove() method cause a problem.

Really strange.