Fixed #617. Proved the issue #617 and fixed it.
Adds now always the `Engine` vendor information in `MysqlSchemaParser` no matter if the build property `addVendorInfo` is true or not, because we need it in `MysqlPlatform` to check whether the table is `InnoDB` and thus with foreignKey support.
If `addVendorInfo` is true, it extracts **all** vendor information (as before).
Added a test that proves the bug and a helper class for more tests like that.