Declare Peer classes in PHP5ObjectBuilder
Non-base Peer classes are used in generated Base classes, but are not
declared (thus, no use statement). This was never an issue until we had
the base classes in om in a separate namespace from the non-base class
stuff.

This fix adds calls to declareClassFromBuilder() in a few places where
it was previously just being called for Object and Query classes.