Coerce offset and limit values to integers for MySQL LIMIT clause
When constructing a MySQL LIMIT clause, values for the offset and limit are coerced to integers. This prevents arbitrary SQL from being injected via a query limit. Example:

    UserQuery::create()->limit('1;DROP TABLE users')->find();

Previously, this would have injected `DROP TABLE users` into the generated SQL. Now, the limit value would be coerced to the integer `1`.

This is similar to the fix for Propel2 [1].

Fixes #1052

[1] https://github.com/propelorm/Propel2/pull/1464